# docs/api.yaml
openapi: 3.1.0
info:
  title: App API
  version: 0.1.0
  description: OpenAPI specification of App.
servers:
  - url: http://localhost:3080
    description: Local development server
tags:
  - name: Authentication
  - name: Account
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      operationId: loginOrCreateAccount
      summary: Trigger login or account creation
      description: Triggers a confirmation email containing a temporary 6-digit verification code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: john@doe.com
      responses:
        '204':
          description: Verification email sent.
        default:
          $ref: '#/components/responses/Error'
  /auth/verify:
    post:
      tags:
        - Authentication
      operationId: verifyAuthenticationCode
      summary: Verify email code
      description: Verifies the provided email and 6-digit code, then creates a new session. If no account exists for the email address, a new account is created.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
            example:
              email: john@doe.com
              otp: "123456"
      responses:
        '200':
          description: Verified account and created session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
              example:
                access_token: eyJhbGciOa...
                refresh_token: X4fFxvY89ypw5f...
                expires_in: 900
                refresh_token_expires_in: 5184000
        default:
          $ref: '#/components/responses/Error'
  /auth/refresh:
    post:
      tags:
        - Authentication
      operationId: refreshSession
      summary: Refresh session tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            example:
              refresh_token: X4fFxvY89ypw5f...
      responses:
        '200':
          description: Refreshed tokens.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
              example:
                access_token: eyJhbGciOiJIU...
                refresh_token: eyJhbGciOiJIU...
                expires_in: 900
                refresh_token_expires_in: 5184000
        default:
          $ref: '#/components/responses/Error'
  /account:
    get:
      tags:
        - Account
      operationId: getAccount
      summary: Get account details
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Account information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
              example:
                email: john@doe.com
                created_at: 1777872664
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags:
        - Account
      operationId: deleteAccount
      summary: Delete account
      description: Deletes the authenticated account.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Account deleted.
        default:
          $ref: '#/components/responses/Error'
  /account/email/change:
    patch:
      tags:
        - Account
      operationId: requestEmailChange
      summary: Request an email address change
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeEmailRequest'
            example:
              new_email: new@email.com
      responses:
        '204':
          description: Email change request accepted.
        default:
          $ref: '#/components/responses/Error'
  /account/email/change/confirm:
    patch:
      tags:
        - Account
      operationId: confirmEmailChange
      summary: Confirm an email address change
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeEmailConfirmRequest'
            example:
              otp: "123456"
      responses:
        '204':
          description: Email change confirmed.
        default:
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
      properties:
        email:
          type: string
          format: email
    VerifyRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - otp
      properties:
        email:
          type: string
          format: email
        otp:
          type: string
          description: Six-digit one-time password.
          minLength: 6
          maxLength: 6
          pattern: '^[0-9]{6}$'
    RefreshRequest:
      type: object
      additionalProperties: false
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    SessionTokens:
      type: object
      additionalProperties: false
      required:
        - access_token
        - refresh_token
        - expires_in
        - refresh_token_expires_in
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          minimum: 0
        refresh_token_expires_in:
          type: integer
          minimum: 0
    VerifyResponse:
      $ref: '#/components/schemas/SessionTokens'
    RefreshResponse:
      $ref: '#/components/schemas/SessionTokens'
    Account:
      type: object
      additionalProperties: false
      required:
        - email
        - created_at
      properties:
        email:
          type: string
          format: email
        created_at:
          type: integer
          description: Account creation timestamp.
    ChangeEmailRequest:
      type: object
      additionalProperties: false
      required:
        - new_email
      properties:
        new_email:
          type: string
          format: email
    ChangeEmailConfirmRequest:
      type: object
      additionalProperties: false
      required:
        - otp
      properties:
        otp:
          type: string
          description: Six-digit one-time password.
          minLength: 6
          maxLength: 6
          pattern: '^[0-9]{6}$'
    Error:
      type: object
      description: Generic JSON error payload.
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
          description: Localized human-readable error message.